Refactor search page into smaller components
